Added resumes page and icons
