+++ app/(website)/menu-items.ts
@@ -1,6 +1,5 @@
 export const menuItems: { href: string; name: string; new?: boolean }[] = [
   { href: "/", name: "About" },
-  // {href: "/resume", name: "Resume"},
-  // {href: "/project", name: "Project"},
+  { href: "/resume", name: "Resume" },
   { href: "/blog", name: "Blog" },
 ];
